+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.5
+
+* Added `memCacheWidth` and `memCacheHeight` to `OmniImage` and `OmniNetworkImage`.
+* Forwarded memory cache resize options to `CachedNetworkImage`.
+* Updated README with guidance and examples for memory-safe caching of large images.
+
 ## 0.0.1
 
 * TODO: Describe initial release.